move the destination functions from db to destinations
